Peut etre ?

+++ StudentLibrary/Solution1.cpp
@@ -19,35 +19,35 @@ float Solution1::ConvertTemperature(float _value, TemperatureUnits _from, Temper
 			if (_to == TemperatureUnits::FAHRENHEIT)
 			{
 				resultat = ((9 / 5) * _value) + 32;
-				return std::round(resultat * 100) / 100;
+				return std::roundf(resultat * 100) / 100;
 			}
 			else if (_to == TemperatureUnits::KELVIN)
 			{
 				resultat = _value + 273.15;
-				resultat = std::round(resultat * 100) / 100;
+				resultat = std::roundf(resultat * 100) / 100;
 				return resultat;
 			}
 		case TemperatureUnits::FAHRENHEIT:
 			if (_to == TemperatureUnits::CELSIUS)
 			{
 				resultat = (5 / 9) * (_value - 32);
-				return std::round(resultat * 100) / 100;
+				return std::roundf(resultat * 100) / 100;
 			}
 			else if (_to == TemperatureUnits::KELVIN)
 			{
 				resultat = ((5 / 9) * (_value - 32)) + 273.15;
-				return std::round(resultat * 100) / 100;
+				return std::roundf(resultat * 100) / 100;
 			}
 		case TemperatureUnits::KELVIN:
 			if (_to == TemperatureUnits::FAHRENHEIT)
 			{
 				resultat = ((9 / 5) * (_value - 273.15)) + 32;
-				return std::round(resultat * 100) / 100;
+				return std::roundf(resultat * 100) / 100;
 			}
 			else if (_to == TemperatureUnits::CELSIUS)
 			{
 				resultat = _value - 273.15;
-				return std::round(resultat * 100) / 100;
+				return std::roundf(resultat * 100) / 100;
 			}
 		default:
 			break;
